scripts/scrapers/requirements.txt
supabase>=2.3.0
google-generativeai
